Move TimeTag and TimeConstants from data to util package
