+++ arktwin/common/src/main/scala/arktwin/common/data/DurationExtensions.scala
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2024-2025 TOYOTA MOTOR CORPORATION
 package arktwin.common.data
+import arktwin.common.util.{MachineTag, VirtualTag}
 
 object DurationExtensions:
   extension (a: Duration)
