+++ arktwin/common/src/main/scala/arktwin/common/util/TimeConstants.scala
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2024-2025 TOYOTA MOTOR CORPORATION
-package arktwin.common.data
+package arktwin.common.util
 
 object TimeConstants:
   inline val nanosPerSecond = 1_000_000_000L
