+++ arktwin/common/src/test/scala/arktwin/common/data/TaggedTimestampSpec.scala
@@ -2,6 +2,7 @@
 // Copyright 2024-2025 TOYOTA MOTOR CORPORATION
 package arktwin.common.data
 
+import arktwin.common.util.{MachineTag, VirtualTag}
 import com.github.plokhotnyuk.jsoniter_scala.core.*
 import org.scalatest.funspec.AnyFunSpec
