+++ arktwin/edge/src/main/scala/arktwin/edge/actors/sinks/Chart.scala
@@ -3,12 +3,12 @@
 package arktwin.edge.actors.sinks
 
 import arktwin.center.services.ChartAgent
+import arktwin.common.data.TaggedTimestamp
 import arktwin.common.data.TimestampExtensions.*
 import arktwin.common.data.Vector3EnuExtensions.*
-import arktwin.common.data.{TaggedTimestamp, VirtualTag}
 import arktwin.common.util.BehaviorsExtensions.*
 import arktwin.common.util.CommonMessages.Nop
-import arktwin.common.util.MailboxConfig
+import arktwin.common.util.{MailboxConfig, VirtualTag}
 import arktwin.edge.actors.EdgeConfigurator
 import arktwin.edge.configs.CullingConfig
 import org.apache.pekko.actor.typed.SpawnProtocol.Spawn
